move jwt and logger to services
